Move vIRQ passthrough handling to architecture independent code
