+++ examples/simple/vmm.c
@@ -75,15 +75,6 @@ extern char _guest_initrd_image_end[];
 /* Microkit will set this variable to the start of the guest RAM memory region. */
 uintptr_t guest_ram_vaddr;
 
-static void serial_ack(size_t vcpu_id, int irq, void *cookie)
-{
-    /*
-     * For now we by default simply ack the serial IRQ, we have not
-     * come across a case yet where more than this needs to be done.
-     */
-    microkit_irq_ack(SERIAL_IRQ_CH);
-}
-
 void init(void)
 {
     /* Initialise the VMM, the VCPU(s), and start the guest */
@@ -113,7 +104,7 @@ void init(void)
         LOG_VMM_ERR("Failed to initialise emulated interrupt controller\n");
         return;
     }
-    success = virq_register(GUEST_VCPU_ID, SERIAL_IRQ, &serial_ack, NULL);
+    success = virq_register_passthrough(GUEST_VCPU_ID, SERIAL_IRQ, SERIAL_IRQ_CH);
     assert(success);
     /* Just in case there is already an interrupt available to handle, we ack it here. */
     microkit_irq_ack(SERIAL_IRQ_CH);
@@ -125,7 +116,7 @@ void notified(microkit_channel ch)
 {
     switch (ch) {
     case SERIAL_IRQ_CH: {
-        bool success = virq_inject(0, SERIAL_IRQ);
+        bool success = virq_handle_passthrough(SERIAL_IRQ_CH);
         if (!success) {
             LOG_VMM_ERR("IRQ %d dropped on vCPU %d\n", SERIAL_IRQ, GUEST_VCPU_ID);
         }
