+++ vmm.mk
@@ -63,7 +63,8 @@ ARCH_INDEP_FILES := src/util/printf.c \
 		    src/guest.c \
 		    src/fault.c \
 		    src/linux.c \
-		    src/dtb.c
+		    src/dtb.c \
+		    src/virq.c
 
 CFILES := ${ARCH_INDEP_FILES}
